initial pass at new cloners and data directory reorganization
